move segmenttype definition, add functions to manipulate peer segment list
